+++ lib/farmbot/jwt.ex
@@ -1,0 +1,6 @@
+defmodule Token do
+  defstruct []
+  def create(thing) do
+    {:ok, thing}
+  end
+end
