+++ lib/farmbot/rpc/transport/gen_mqtt/gen_mqtt_client.ex
@@ -1,5 +1,5 @@
 defmodule Farmbot.RPC.Transport.GenMqtt.Client do
-  use Syncables
+   
   @moduledoc """
     MQTT transport for farmbot RPC Commands.
   """
